Add templates + CSS for ATF's posters
ATF has two poster PDFs which we would like to include right before the
section that they are describing. We recently added the ability to add
arbitrary, custom templates as well as some fixups to paragraph markers and
a hook for agency-specific modules. All combined, they allow us to write a
shared template `poster_wrapper.html`, with some custom CSS, and include that
at the two paragraphs.