First draft of docker-for-development rework.
This replaces the existing docker-compose file with one that will account for
local edits more easily. As an innovation over omb-eregs, this places the
Python and Node libraries in a named volume rather than using the current dir
(H/T @toolness).

This should make local development much easier as it doesn't require setting
up all of the Python and Node dependencies. This doesn't account for the
parser just yet.