Clear copied frontend *before* collectstatic.
Django will look first in the `compiled` directory (then in the
regulations-site dirs) when searching for a static file. This is the correct
behavior when running the application, but not when we're building the
frontend (i.e. when we're filing that `compiled` directory). We should clear
out `compiled` earlier in the build script to accommodate this need.