Clone ATF-ruling logic from parser into API.
The `atf-rulings` layer has thus far been generated with a parser plugin.
We'd like to re-process the data *without* needing to host the parser; our
current strategy for doing that is to skip the parser and write data directly
to the API. This implements that by copying (and rewriting) many of the
functions present in the parser plugin.