Always attach postgres.
Currently, we only attach postgres to the running ATF instance if we run
`dev-with-db` (otherwise, we're running with the live ATF data, skipping the
SQLite-based local API). As we test a postgres-based search, we need to have
Postgres running (even if it's not actively used).

This accomplishes that by using a specific env variable to signify that we
want to use the live data (rather than just testing for a database
connection).