Allow atf_resources to handle multiple doc types
Previously, we had more or less ported the "Rulings" logic over from the
parser. Now that we have a better idea of how the data'll be structured, mock
it out in a YAML file. This replaces the previous logic entirely, so probably
easiest to review by looking only at the new code.

This also includes some manually-generated data.