Tweak frontend build process.
Previously, we had a bunch of shell commands executing in the host which would
combine files from -site and ATF before they could be built. This doesn't play
nice with Circle, which doesn't synchronize file access in Docker containers
with the host. Instead, let's perform all actions within Docker containers.

As an upshot, this also makes running on Windows easier.