Initial redeployer micro app.
To satisfy ATO requirements, we generally need to redeploy on a frequent
basis. In the past, we've done this via Travis's daily cron job, but we can't
rely on that going forward, so we need to write a little app that does the
deploys for us.

This script will wait until the requested time (or date), then hit the
CircleCI API requesting a rebuild, then wait until the next time. Generally, a
deployment will redeploy the script, too, so that second nap will likely be
interrupted.