Rework cloud.gov manifest to include redeployer app.
As the env needs of the redployer app are rather different from the web app,
we also move those services/env keys into their respective apps.

We're bumping up against our memory limit when deploying, so this grabs a
small amount of memory from the worker app to run the redeployer.